Refactor: Move adam_step tests to Catch2 framework
